move weather api and weather api client to remote.
