+++ src/main/kotlin/data/remote/WeatherApiClient.kt
@@ -1,4 +1,4 @@
-package data.remote.repository
+package data.remote
 
 import io.ktor.client.*
 import io.ktor.client.call.*
