+++ src/main/kotlin/data/repository/WeatherRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package data.repository
 
 import data.mapper.toWeather
-import data.remote.repository.WeatherApi
+import data.remote.WeatherApi
 import logic.model.Weather
 import logic.repository.WeatherRepository
